# INSTA-LITE/insta-lite-back/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginRequest:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: hugues@gmail.com
        type: string
      fullName:
        example: Goli yao Hugues
        type: string
      password:
        example: pass1234
        minLength: 6
        type: string
      username:
        example: huguesgoli
        minLength: 3
        type: string
    required:
    - email
    - fullName
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: huguescodeur@gmail.com
    name: Hugues Codeur
  description: API REST pour une application de type Instagram (posts, likes, commentaires,
    follow).
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Insta-Lite API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authentifie un utilisateur (email ou username) et retourne le token
        JWT
      parameters:
      - description: Identifiants de connexion
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON invalide ou validation échouée
          schema:
            type: string
        "401":
          description: Identifiants incorrects
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      summary: Connexion
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Crée un compte et retourne le token JWT
      parameters:
      - description: Données d'inscription
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON invalide ou validation échouée
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      summary: Inscription d'un nouvel utilisateur
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Entrez le token JWT avec le préfixe "Bearer ". Exemple: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"